Loader: call BmSetMemoryTypeInformationVariable
Note that BmSetMemoryTypeInformationVariable() is told by the loader that the image being loaded is *not* a boot application, which is a blatant lie. The reason for doing this is to prevent BmSetMemoryTypeInformationVariable() from performing a warm reset of the system due to a change in the memory type info. (Again, and again, and again...)

This is an attempt to fix S4 (hibernate) entry/resume issues; see #12.