Move BacktrackToFunctionStart to pe.c
This function is highly PE-specific, and pe.c is therefore a more suitable TU.
Also rename it to FindFunctionStart, so as to not imply that any state (e.g. the virtual runtime address used for decoding, ..) is modified, nor that we necessarily need to be backtracking when calling it. This function can also be used more generally to verify whether an address is part of any function at all according to the PE exception directory.

Misc: minor const correctness fixes