Increase file and section alignments to equal page size
Some environments restrict images from running if sections are not page-aligned, because this prevents them from fully enforcing section protections such as RO and NX in hardware via page tables.

Also increase the file alignment to match the new section alignment, because identical file and section alignments are a requirement for running images as XIP. XIP execution is not the norm for EfiGuard, but it at least shouldn't be prevented over a few KBs of difference in binary file size.